requirements.txt
# =====
# Dynamic Loan Pricing Engine - Dependencies
# =====

# --- Core Data Processing ---
pandas>=2.0.0
numpy>=1.24.0

# --- Machine Learning ---
scikit-learn>=1.3.0
xgboost>=2.0.0
optuna>=3.0.0

# --- Explainability ---
shap>=0.43.0

# --- Database ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# --- Optimization (Phase 2+) ---
scipy>=1.11.0

# --- Dashboard (Phase 2+) ---
streamlit>=1.30.0

# --- Model Serialization ---
joblib>=1.3.0

# --- Configuration ---
pyyaml>=6.0
python-dotenv>=1.0.0

# --- Visualization ---
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# --- Utilities ---
tqdm>=4.65.0

# --- Agentic Layer (Phase 3/4) ---
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
tavily-python>=0.5.0

# --- API Layer (Phase 3/4) ---
fastapi>=0.110.0
uvicorn>=0.27.0

# --- Scheduler (Phase 3/4) ---
apscheduler>=3.10.0
